added .dll
Getting there ive added a dll to a C# projekt that can extract the code we need. Now we have a issue that it says start is not a function it might have something to do with time, so we might need to wait, the C# part is working on its own. Now we need to integrate it correctly in the node projekt. The parts are working in the Node project but the connection is off.